+++ app/src/main/java/com/example/sampleappbyme/main/util/AddEditTaskNavigator.kt
@@ -1,4 +1,4 @@
-package com.example.sampleappbyme.main
+package com.example.sampleappbyme.main.util
 
 interface AddEditTaskNavigator {
